% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud}
\alias{raincloud}
\title{raincloud}
\usage{
raincloud(data, mapping, flip_coords = FALSE, ...)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{flip_coords}{Default orientation is vertical (FALSE),
set to TRUE for horizontal.}

\item{...}{Additional parameters.}
}
\value{
Returns a raincloud ggplot.
}
\description{
ggplot2 raincloud plot template.
}
\examples{
# Example usage of the function
plot <- raincloud(iris, aes(Species, Sepal.Width, color = Species))

}
