% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_barcode.R
\name{raincloud_barcode}
\alias{raincloud_barcode}
\title{raincloud_barcode}
\usage{
raincloud_barcode(data, mapping, barcode_size = 15, flip_coords = FALSE, ...)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{barcode_size}{Size of barcode marks; defaults to 15.}

\item{flip_coords}{Default orientation is vertical (FALSE),
set to TRUE for horizontal.}

\item{...}{Additional parameters.}
}
\value{
Returns a raincloud ggplot.
}
\description{
ggplot2 raincloud plot template with a barcode instead of points.
}
\examples{
# Example usage of the function
plot <- raincloud_barcode(iris, ggplot2::aes(Species, Sepal.Width))

}
