% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_boxjitter.R
\name{raincloud_boxjitter}
\alias{raincloud_boxjitter}
\title{raincloud_boxjitter}
\usage{
raincloud_boxjitter(data, mapping, flip_coords = FALSE, ...)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{flip_coords}{Default orientation is vertical (FALSE),
set to TRUE for horizontal.}

\item{...}{Additional parameters.}
}
\value{
Returns a raincloud ggplot.
}
\description{
ggplot2 raincloud plot template with the "rain" inside the boxplot.
}
\examples{
# Example usage of the function
plot <- raincloud_boxjitter(iris, ggplot2::aes(Species, Sepal.Width))

}
