% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_raincloud.R
\name{theme_raincloud}
\alias{theme_raincloud}
\title{theme_raincloud}
\usage{
theme_raincloud(flip_coords = FALSE, ...)
}
\arguments{
\item{flip_coords}{Default orientation is vertical (FALSE),
set to TRUE for horizontal.}

\item{...}{Additional parameters.}
}
\value{
Returns a raincloud ggplot.
}
\description{
ggplot2 raincloud plot theme based on ggplot2::theme_minimal().
}
\examples{
# Example usage of the function
plot <- raincloud(iris, ggplot2::aes(Species, Sepal.Width)) +
theme_raincloud()

}
